fixed template issue in glue

+++ src/python/lib/graph/lifted_multicut/learnable_lifted_multicut_objective.cxx
@@ -20,8 +20,8 @@ namespace lifted_multicut{
     void exportLearnableLiftedMulticutObjectiveT(py::module & liftedMulticutModule) {
 
         typedef GRAPH Graph;
-        typedef LiftedMulticutObjective<Graph, float> BaseObjectiveType;
-        typedef LearnableLiftedMulticutObjective<Graph, float> ObjectiveType;
+        typedef LiftedMulticutObjective<Graph, double> BaseObjectiveType;
+        typedef LearnableLiftedMulticutObjective<Graph, double> ObjectiveType;
         typedef typename ObjectiveType::LiftedGraphType LiftedGraphType;
         const auto clsName = LiftedMulticutObjectiveName<ObjectiveType>::name();
